fix links and link shorteners
goog.le and bit.ly now seem to enforce a max length on the urls that they'll shorten. infuriating.

this changes behavior to save the key instead of the graphState. We found that people weren't changing the graph state (nor was it even clear that you could / should). Fixes link shorteners. bummer.

Ideal next behavior: allow you to import a state from desmos. We'd then store the graph hash and load the graph through Ajax based on the hash. That's better than storing it here anyhow, because that graph becomes portable.